(1248) Split up shantytownActorModel
